HDMI: update mode selection logic
The kernel now passes all of the modes to user space with flags
to specify whether or not it is preferred or not and if the
hardware actually supports the mode. With this, we can try to
pick modes that match the 'preferred' mode as specified in the EDID
from the panel.

If the preferred mode is masked due to the driver, the logic will
try to favor available modes that match the resolution of the preferred
mode. If none of these options are available, fall back to the
older logic of trying to find the best mode to match what the
current framebuffer is set to.

(cherry picked from commit c63364590d5df4837438ba69b4ba7dad66af3321)

Conflicts:

	hwc/hwc.c

Change-Id: Ic796819b60665ec34f4c9967f31f0cc1f2cb7a28