HWC: OMAP4: disable the first frame display reset
A hack to reset the display on the first frame was put in previously
to allow for a seamless transition between the boot logo from
the bootloader to the boot animation.

In our case, the kernel/hwc has already scanned and set the proper
hdmi mode, so we dont want to bounce the panel.

(cherry picked from commit eafa266c88c496aa539f46407fbdab137b86d2c7)

Change-Id: I858185c8a8416bff84ee77f67100c25f754c16b9