hwc: rgz: Fix incorrect src rectangle calculation for scaled layers
When scaling a layer, the blit is programmed in a way where the src
rectangle uses the whole cropping rectangle coordinates and
dimensions. However if the layer has a transform, the width and the
height need to be swapped not only in the geometry but also in the
src rectangle to match the new dimensions.

This patch adds the missing width and height swapping for the
src rectangle if the transformation is 90 or 270 degrees.

Change-Id: Ia9d822b6698210573fff662f2fcb129dfc048bf7
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>
(cherry picked from commit a80084da8fb2c63a2e44ef9d4ce7598b7cad853f)