Fake VSYNC events if needed.
If HWC clients (aka, surface flinger) want to be receiving vsync
events, and HDMI is unplugged, fake the events instead.  This is
particularly important at startup on JB; the system thinks that it has
turned the display on and enabled VSYNC events, failure to deliver
them results in Bad Things.  We cannot rely on HDMI hardware to
delvier the events if the HDMI link is not established, so we just
fake them instead.

(cherry-picked from 25ab304446afba89283c1a7c7225a69cc3cf4385)

Change-Id: Id4806f96d37cf8ad5b6a05d9077b3d88fd24f4d5
Signed-off-by: John Grossman <johngro@google.com>

Conflicts:
	hwc/hwc.c