android: binder: reduce memory for binder debugfs
Reduce the amount of memory the binder debugfs seq_file's need
to output the large amount of data they provide.  This breaks up
the individual reads into smaller objects to limit the kmalloc'd
memory to a PAGE.  This helps reduce OOM conditions when these
files are read in mildly fragmented systems.