Rename ARM assembler push/pull macros
The ARM assembler library functions use a macro called push that
along with a macro called pull is used to shift bytes around in a
word in an endian-independent way. However, the modern unified
ARM assembler syntax also defines the instruction push to push
data onto the stack, which has specific encodings in the Thumb2
instruction set. For prevent possible conflicts going forward, and
to allow the use of the more transparent push instruction along
with the modern unified assembler syntax, this patch renames all
occurrences of the push macro to pushbits, as well as renaming
the macro argument, when also called push, to pushshift.

For consistency, the macro called pull with its argument name
pull are also renamed to pullbits and pullshift, respectively.